+++ apm-sniffer/apm-agent-core/src/main/java/org/apache/skywalking/apm/agent/core/util/InstanceJsonPropertiesUtil.java
@@ -20,24 +20,15 @@ package org.apache.skywalking.apm.agent.core.util;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
-
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
-
 import org.apache.skywalking.apm.agent.core.conf.Config;
-import org.apache.skywalking.apm.agent.core.logging.api.ILog;
-import org.apache.skywalking.apm.agent.core.logging.api.LogManager;
+import org.apache.skywalking.apm.agent.core.version.Version;
 import org.apache.skywalking.apm.network.common.v3.KeyStringValuePair;
 import org.apache.skywalking.apm.util.StringUtil;
 
 public class InstanceJsonPropertiesUtil {
-
-    private static final ILog LOGGER = LogManager.getLogger(InstanceJsonPropertiesUtil.class);
-    private static final String GIT_PROPERTIES = "skywalking-agent-git.properties";
     private static final Gson GSON = new Gson();
 
     public static List<KeyStringValuePair> parseProperties() {
@@ -55,32 +46,8 @@ public class InstanceJsonPropertiesUtil {
 
         properties.add(KeyStringValuePair.newBuilder().setKey("namespace").setValue(Config.Agent.NAMESPACE).build());
         properties.add(KeyStringValuePair.newBuilder().setKey("cluster").setValue(Config.Agent.CLUSTER).build());
-        properties.add(KeyStringValuePair.newBuilder().setKey("version").setValue(getAgentVersion()).build());
+        properties.add(KeyStringValuePair.newBuilder().setKey("version").setValue(Version.CURRENT.toString()).build());
 
         return properties;
-    }
-
-    public static String getAgentVersion() {
-        try {
-            InputStream inStream = InstanceJsonPropertiesUtil.class.getClassLoader()
-                    .getResourceAsStream(GIT_PROPERTIES);
-            if (inStream != null) {
-                Properties gitProperties = new Properties();
-                gitProperties.load(inStream);
-                String commitIdAbbrev = gitProperties.getProperty("git.commit.id.abbrev");
-                String buildTime = gitProperties.getProperty("git.build.time");
-                String buildVersion = gitProperties.getProperty("git.build.version");
-                String version =  buildVersion + "-" + commitIdAbbrev + "-" + buildTime;
-
-                LOGGER.info("SkyWalking agent version: {}", version);
-                return version;
-            } else {
-                LOGGER.warn("{} not found, SkyWalking agent version: unknown", GIT_PROPERTIES);
-            }
-        } catch (IOException e) {
-            LOGGER.error("Failed to get agent version", e);
-        }
-
-        return "UNKNOWN";
     }
 }