+++ tools/releasing/create_release.sh
@@ -59,6 +59,11 @@ git checkout ${TAG_NAME}
 git submodule init
 git submodule update
 
+# Generate a static version.properties and override the template when releasing source tar
+# because after that there is no Git information anymore.
+./mvnw -pl apm-sniffer/apm-agent-core initialize \
+       -DgenerateGitPropertiesFilename="$(pwd)/apm-sniffer/apm-agent-core/src/main/resources/version.properties"
+
 cd ..
 # Build source code tar
 tar czf ${PRODUCT_NAME}-src.tgz \